Move internal components to private dir
