+++ projects/js-packages/charts/src/components/private/chart-composition/chart-html.tsx
