+++ projects/js-packages/charts/src/components/private/chart-composition/chart-svg.tsx
