+++ projects/js-packages/charts/src/components/private/chart-composition/index.ts
