+++ projects/js-packages/charts/src/components/private/chart-composition/types.ts
@@ -1,4 +1,4 @@
-import { Legend } from '../legend';
+import { Legend } from '../../legend';
 import type { ComponentType, FC, PropsWithChildren } from 'react';
 
 /**
