+++ projects/js-packages/charts/src/components/private/default-glyph/index.ts
