+++ projects/js-packages/charts/src/components/private/grid-control/grid-control.module.scss
