+++ projects/js-packages/charts/src/components/private/grid-control/stories/index.stories.tsx
