+++ projects/js-packages/charts/src/components/private/with-responsive/index.ts
